new db interface

+++ TDDMZ/TDCashierViewController.m
@@ -80,7 +80,7 @@
         count += good.goods_number * [good.shop_price integerValue];
     }
     
-    money = [NSString stringWithFormat:@"%ld",count];
+    money = [NSString stringWithFormat:@"%ld",(long)count];
     
     return money;
 }
